Move API key code to api/ package
This is because it is useful for low-level consumers.
