+++ api/testdata/.hasspaceinit.key
